Add theme chossing button and rid of card count adjuster legos
